Done: Refactoring OpenDonationDetailedDescriptorView
